+++ commands/publish/__tests__/publish-from-git.test.js
@@ -114,26 +114,20 @@ describe("publish from-git", () => {
     });
 
     const cwd = await initFixture("normal");
+    const command = lernaPublish(cwd)("from-git");
 
-    try {
-      await lernaPublish(cwd)("from-git");
-    } catch (err) {
-      expect(err.message).toBe("uncommitted");
-      // notably different than the actual message, but good enough here
-    }
-
-    expect.assertions(1);
+    await expect(command).rejects.toThrow("uncommitted");
+    // notably different than the actual message, but good enough here
   });
 
   it("throws an error when --git-head is passed", async () => {
     const cwd = await initFixture("normal");
+    const command = lernaPublish(cwd)("from-git", "--git-head", "deadbeef");
 
-    try {
-      await lernaPublish(cwd)("from-git", "--git-head", "deadbeef");
-    } catch (err) {
-      expect(err.prefix).toBe("EGITHEAD");
-    }
-
-    expect.hasAssertions();
+    await expect(command).rejects.toThrow(
+      expect.objectContaining({
+        prefix: "EGITHEAD",
+      })
+    );
   });
 });