+++ commands/version/__tests__/get-current-branch.test.js
@@ -11,10 +11,6 @@ test("getCurrentBranch", async () => {
 
 test("getCurrentBranch without commit", async () => {
   const cwd = await initFixture("root-manifest-only", false);
-  expect.assertions(1);
-  try {
-    await getCurrentBranch({ cwd });
-  } catch (err) {
-    expect(err.message).toMatch(/Command failed: git rev-parse --abbrev-ref HEAD.*/);
-  }
+
+  expect(() => getCurrentBranch({ cwd })).toThrow(/Command failed: git rev-parse --abbrev-ref HEAD.*/);
 });